backend/requirements.txt
livekit-agents>=0.8.0
livekit>=1.0.0
livekit-plugins-google
python-dotenv
google-generativeai
mem0ai
chromadb
faster-whisper
edge-tts
pydub
flask
flask-cors
AppOpener
pure-python-adb
opencv-python
pillow
openwakeword
langdetect
webrtcvad-wheels
requests
